releaseagent: remove unused release discovery code
